requirements.txt
yamale >= 4.0.4
PyYAML >= 6.0
ruamel.yaml >=  0.17.21
GitPython >= 3.1.29
python-terraform >= 0.10.1
git-url-parse >= 1.2.2
graphviz >= 0.19.0
diagrams >= 0.23.1